:recycle: price 필드 Long type 으로 변경

+++ src/main/java/com/potato/ecommerce/domain/product/dto/ProductUpdateRequest.java
@@ -18,7 +18,7 @@ public class ProductUpdateRequest {
     private String description;
 
     @NotBlank
-    private int price;
+    private Long price;
 
     @NotBlank
     private int stock;
